moves satellite id route out of home
