+++ views/homepage.handlebars
@@ -3,46 +3,43 @@
     {{#if loggedIn}}
 
     {{#each satellites as |satellite|}}
+    <a href="/api/satellite/{{satellite.id}}">
+        <section class="homepageDiv">
 
-    <section class="homepageDiv">
 
-        <a href="/{{satellite.id}}">
             <div class="satellite" {{satellite.norad_id}}>
                 <h2>{{satellite.satellite_name}}</h2>
             </div>
             <div class="country">
                 <p>Created by: {{satellite.country.country_name}}</p>
             </div>
-        </a>
-        <div>
-            <button>Remove from View</button>
-        </div>
 
-    </section>
+            <div>
+                <button>Remove from View</button>
+            </div>
 
+        </section>
+    </a>
 
     {{/each}}
 
     {{else}}
 
     {{#each satellites as |satellite|}}
+    <a href="/api/satellite/{{satellite.id}}">
+        <section class="homepageDiv">
 
-    <section class="homepageDiv">
 
-        <a href="/{{satellite.id}}">
             <div class="satellite" {{satellite.norad_id}}>
                 <h2>{{satellite.satellite_name}}</h2>
             </div>
             <div class="country">
                 <p>Created by: {{satellite.country.country_name}}</p>
             </div>
-        </a>
-        <div>
-            <button>Remove from View</button>
-        </div>
 
-    </section>
 
+        </section>
+    </a>
     {{/each}}
 
     {{/if}}
